Create README.md
Task is to create a To-Do List Application in Python. This application should allow users to add tasks, mark tasks as completed, update task descriptions, and remove tasks from the list. You can choose whether to make it a command-line application or implement a graphical user interface (GUI) for added challenge.